fixed tests before functionality testing

+++ visualswarm/app.py
@@ -110,7 +110,8 @@ def start_application(with_control=False):
         behavior_proc.start()
         motor_control.start()
         system_monitor_proc.start()
-        emergency_proc.start()
+        if with_control:
+            emergency_proc.start()
 
         # Wait for subprocesses in main process to terminate
         visualizer.join()
@@ -121,7 +122,8 @@ def start_application(with_control=False):
         behavior_proc.join()
         motor_control.join()
         system_monitor_proc.join()
-        emergency_proc.join()
+        if with_control:
+            emergency_proc.join()
 
     except KeyboardInterrupt:
         # suppressing all error messages during graceful exit that come from intermingled queues
@@ -130,8 +132,10 @@ def start_application(with_control=False):
         logger.info(f'{bcolors.WARNING}EXIT gracefully on KeyboardInterrupt{bcolors.ENDC}')
 
         # Terminating Processes
-        emergency_proc.terminate()
-        emergency_proc.join()
+        if with_control:
+            emergency_proc.terminate()
+            emergency_proc.join()
+            logger.info(f'{bcolors.WARNING}TERMINATED{bcolors.ENDC} emergency process and joined!')
         system_monitor_proc.terminate()
         system_monitor_proc.join()
         logger.info(f'{bcolors.WARNING}TERMINATED{bcolors.ENDC} system monitor process and joined!')
